+++ data/EFMR/scripts/autoexec.lua
@@ -4,5 +4,7 @@ run_script("section.lua");
 
 run_script("soledade_barra_helix_monitor.lua");
 
+run_script("soledade_barra_monitor.lua")
+
 log_info("Autoexec finished");