+++ src/Emulator/main.cpp
@@ -139,7 +139,8 @@ int main(int, char **)
 	//ArduinoLib::SetSerialInput("/sv;");
 #endif
 
-	//ArduinoLib::SetSerialInput("/cfg TestDevice 242.69.116.44.41.93 7202 8989;sv;");
+	ArduinoLib::SetSerialInput("/cfg TestDevice 242.69.116.44.41.93 7202 8989;sv;");
+	//ArduinoLib::SetSerialInput("/cfg HadesExtA 242.69.116.44.41.93 7202 8989;sv;");
 
 	dcclite::Clock clock;